+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## Release 0.9.1
+## Release 0.9.2
+
+* Added acceptance tests using Puppet Litmus
+
+## Release 0.9.1 (not released on Puppet Forge)
 
 * Added a new parameter to make the ssh key base directory permissions configurable
 * Added more unit test
-* Added acceptance tests using Puppet Litmus
 
 ## Release 0.9.0